+++ ios/BookProgressTracker/AppDelegate.mm
@@ -1,14 +1,11 @@
 #import "AppDelegate.h"
 
 #import <React/RCTBundleURLProvider.h>
-#import <Firebase.h>
 
 @implementation AppDelegate
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  [FIRApp configure];
-  
   self.moduleName = @"BookProgressTracker";
   self.initialProps = @{};